requirements.txt
streamlit
tensorflow
keras
pillow
numpy
python-dotenv
google-generativeai
